+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/factory/SortButtonFactory.java
@@ -1,8 +1,44 @@
 package com.nemonotfound.nemos.inventory.sorting.factory;
 
-public abstract class SortButtonFactory implements ButtonCreator {
+import com.nemonotfound.nemos.inventory.sorting.gui.components.buttons.AbstractInventoryButton;
+import com.nemonotfound.nemos.inventory.sorting.gui.components.buttons.SortButton;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.inventory.AbstractContainerMenu;
 
-    protected int getLeftPosWithOffset(int leftPos, int offset) {
-        return leftPos + offset;
+public class SortButtonFactory extends ButtonFactory {
+
+    private static SortButtonFactory INSTANCE;
+
+    private SortButtonFactory() {}
+
+    public static SortButtonFactory getInstance() {
+        if (INSTANCE == null) {
+            INSTANCE = new SortButtonFactory();
+        }
+
+        return INSTANCE;
+    }
+
+    @Override
+    public AbstractInventoryButton createButton(
+            int startIndex, int endIndex, int leftPos, int topPos, int xOffset, int yOffset, int width,
+            int height, AbstractContainerMenu menu, boolean isInventoryButton
+    ) {
+        var buttonName = Component.translatable("nemos_inventory_sorting.gui.sortButton");
+        var shiftButtonName = Component.translatable("nemos_inventory_sorting.gui.sortButtonShift");
+        var builder = new AbstractInventoryButton.Builder<>(SortButton.class)
+                .startIndex(startIndex)
+                .endIndex(endIndex)
+                .x(getLeftPosWithOffset(leftPos, xOffset))
+                .y(topPos + yOffset)
+                .xOffset(xOffset)
+                .width(width)
+                .height(height)
+                .buttonName(buttonName)
+                .shiftButtonName(shiftButtonName)
+                .menu(menu)
+                .isInventoryButton(isInventoryButton);
+
+        return builder.build();
     }
 }