+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/gui/components/buttons/SortButton.java
@@ -1,20 +1,19 @@
-package com.nemonotfound.nemos.inventory.sorting.gui.components.buttons.sorting;
+package com.nemonotfound.nemos.inventory.sorting.gui.components.buttons;
 
 import com.nemonotfound.nemos.inventory.sorting.ModKeyMappings;
-import com.nemonotfound.nemos.inventory.sorting.gui.components.buttons.AbstractInventoryButton;
-import com.nemonotfound.nemos.inventory.sorting.service.sorting.AlphabeticallyDescendingSortingService;
 import com.nemonotfound.nemos.inventory.sorting.service.InventoryService;
+import com.nemonotfound.nemos.inventory.sorting.service.SortingService;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.resources.ResourceLocation;
 
 import static com.nemonotfound.nemos.inventory.sorting.Constants.MOD_ID;
 
-public class SortAlphabeticallyDescendingButton extends AbstractInventoryButton {
+public class SortButton extends AbstractInventoryButton {
 
-    private final ResourceLocation buttonTexture = ResourceLocation.fromNamespaceAndPath(MOD_ID, "sort_button_alphabetically_dec");
-    private final ResourceLocation buttonHoverTexture = ResourceLocation.fromNamespaceAndPath(MOD_ID, "sort_button_alphabetically_dec_highlighted");
+    private final ResourceLocation buttonTexture = ResourceLocation.fromNamespaceAndPath(MOD_ID, "sort_button");
+    private final ResourceLocation buttonHoverTexture = ResourceLocation.fromNamespaceAndPath(MOD_ID, "sort_button_highlighted");
 
-    public SortAlphabeticallyDescendingButton(Builder<SortAlphabeticallyDescendingButton> builder) {
+    public SortButton(Builder<SortButton> builder) {
         super(builder);
     }
 
@@ -31,7 +30,7 @@ public class SortAlphabeticallyDescendingButton extends AbstractInventoryButton 
     @Override
     public void onClick(double mouseX, double mouseY) {
         var inventoryService = InventoryService.getInstance();
-        var sortingService = AlphabeticallyDescendingSortingService.getInstance();
+        var sortingService = SortingService.getInstance();
         var endIndex = inventoryService.calculateEndIndex(isButtonShiftable(), this.endIndex);
 
         inventoryService.handleSorting(sortingService, menu, startIndex, endIndex);